Some more renames for consistency
